Setting mom equal to specific version
The integration testing is pulling in a bad version of the mom
package.  It's actually pulling in a totally different package
that from TestPyPi that has the package name, mom.  This is
introducing a dependency on Python 3.5 which is causing all kinds
of issues.

I will revert this change once I've complete my tests.