Removed non-production stage
The non-production stage, deploy_staging, was removed because
it will prove to be brittle -- TestPyPi will delete projects and
accounts periodically and without warning.

My initial goal was to establish a repeatable process to deploy
new Tags of this SDK to PyPi.  In the future we can invest in
additional CI capabilities.